Use custom ArrowKeyMovementMethod for controlling auto selection
